+++ src/scip/heur_completesol.c
@@ -805,7 +805,7 @@ SCIP_RETCODE applyCompletesol(
    SCIP_CALL( SCIPsetIntParam(subscip, "display/freq", -1) );
 #else
    /* disable statistic timing inside sub SCIP and output to console */
-   SCIP_CALL( SCIPsetIntParam(subscip, "display/verblevel", SCIP_VERBLEVEL_NONE) );
+   SCIP_CALL( SCIPsetIntParam(subscip, "display/verblevel", (int) SCIP_VERBLEVEL_NONE) );
    SCIP_CALL( SCIPsetBoolParam(subscip, "timing/statistictiming", FALSE) );
 #endif
